Added duplicate function linter rule
Summary:
Defines a linter rule `ArcanistDuplicateFunctionDeclarationXHPASTLinterRule` to give a lint error when a function with the same name (function names are case insensitive) is declared multiple times in the same file.

If you try to redeclare a function it will cause a PHP fatal error and as such should raise linter errors.

A lint error will not be raised if a function is defined in a conditional block. For example the following is valid PHP:

```lang=php
if (true) {
  function a() {
    // Implementation
  }
} else {
  function a() {
    // Alternate implementation
  }
}
```

Test Plan: Ran unit tests and manually tested `arc lint` with local php files.

Reviewers: #core_gaf, josh

Reviewed By: #core_gaf, josh

Subscribers: josh, tpham

Differential Revision: https://phabricator.freelancer.com/D28706