Added duplicate method linter rule
Summary:
Defines a linter rule `ArcanistDuplicateMethodDeclarationXHPASTLinterRule` which raises a lint error when a method with the same name is redeclared in the same class definition.

Method names are case insensitive and as such `duplicate()` and `Duplicate()` are still duplicate methods.

This is required since redeclaring a method with cause a PHP fatal error.

Test Plan: Ran unit tests and manual testing with arc lint on a local .php file.

Reviewers: josh, #core_gaf

Reviewed By: josh, #core_gaf

Subscribers: tpham, josh

Differential Revision: https://phabricator.freelancer.com/D28661