Added duplicate property linter rule
Summary:
Added a linter rule `ArcanistDuplicatePropertyDeclarationXHPASTLinterRule` which will raise a lint error if a property with the same name is redeclared in the same class or trait.

Unlike methods and functions, property names are case sensitive and as such `$property` and `$PROPERTY` are not duplicates.

If a property is redeclared it would cause a PHP fatal error.

Test Plan: Ran unit tests and manually tested with `arc lint` on local php files.

Reviewers: tpham, josh, #core_gaf

Reviewed By: tpham, josh, #core_gaf

Subscribers: josh, tpham

Differential Revision: https://phabricator.freelancer.com/D28755