Add configuration options to ArcanistESLintLinter
Summary: Ref T28698. Adds the ability to specify a config and an env file to `eslint` when linting with arcanist. Also adds configuration that makes the unit tests work properly with a specified rule set.

Test Plan: Ran unit tests and tested manually on my local machine with a number of different configurations.

Reviewers: #core_gaf, josh

Reviewed By: #core_gaf, josh

Subscribers: josh

Maniphest Tasks: T28698

Differential Revision: https://phabricator.freelancer.com/D32730