Write a basic Composer linter
Summary: Adds a Composer linter which attempts to verify that the `composer.lock` file is up-to-date. This attempts to catch an issue which I saw earlier today in which a developer manually modified `composer.json` and didn't run `composer update`.

Test Plan: Added whitespace to `composer.json` and saw the linter raise an error.

Reviewers: tpham, raymond.huang

Reviewed By: raymond.huang

Subscribers: michael.foley

Differential Revision: https://phabricator.freelancer.com/D23097