Improve unit tests for `ArcanistDockerContainerLinterProxy`
Summary: Ref T87915. In D125570, I modified `ArcanistDockerContainerLinterProxyTestCase::getLinter()` such that the `ArcanistExternalLinter` being returned is preconfigured to proxy a **mock** linter. This is desirable for //most// of the test cases in `ArcanistDockerContainerLinterProxyTestCase`, but it isn't desirable for the `testLinter()` method, which is executing real commands. To fix this, split the current `getLinter()` method into `getLinter()` and `getLinterWithMockProxiedLinter()`.

Test Plan: Ran unit tests.

Reviewers: tpham, szhu

Reviewed By: tpham

Maniphest Tasks: T87915

Differential Revision: https://phabricator.tools.flnltd.com/D126257