Add 'FlarcJSHintLinter'
Summary: Ref T87915. A problem with `ArcanistJSHintLinter` is that it uses its own `reporter.js` to parse the output of `jshint` to JSON, which requires `ArcanistDockerContainerLinterProxy` to mount the `reporter.js` specifically as a special case. This `FlarcJSHintLinter` uses new feature from JSHint's checkstyle XML output and parse that in PHP instead to eliminate the usage of `reporter.js`

Test Plan: arc unit + test with `gaf`

Reviewers: josh, O8 Phabricator

Reviewed By: josh, O8 Phabricator

Maniphest Tasks: T87915

Differential Revision: https://phabricator.tools.flnltd.com/D126914