Add 'show-info' flag to psalm linter
Summary:
Rules that are set to `info` should be displaying as advice on phabricator and during `arc lint`.
Currently this is not the case because the flag is not added in when the command is being ran.

This diff adds that flag in so that they are properly displayed as advice

Test Plan:
arc lint and get the info to show as lint advice

Reviewers: O8 Phabricator, dwang, #developer_experience, jchen

Reviewed By: O8 Phabricator, dwang, #developer_experience, jchen

Subscribers: jchen

Differential Revision: https://phabricator.tools.flnltd.com/D168055