[flarc] Fix up parsing of config file
Summary:
Was wondering why it was taking so long for my branch to push, it ended up failing to curl the request. This was because it took my token from k8s since it was declared below it.

This diff adds sets the `parse sections` parameter of parse_ini_file to true so it properly processes the different sections so the token doesnt get muddled up with k8s jenkins

Test Plan:
works now for mergequeuepush

```name=arc mergequeue-push D174731
Submitting revision 'D174731: Fix stashing of junit phpunit results in phptest'...
 SUBMIT  Submitting to merge queue push...
Done submitting to mergequeue-push job. Branch: ahe-D174731
```
{F2516917}
https://ci.tools.flnltd.com/job/Developers/job/mergequeue-push/447/

Reviewers: #developer_experience, hzhang

Reviewed By: #developer_experience, hzhang

Differential Revision: https://phabricator.tools.flnltd.com/D174732