Add ArcanistRectorLinter implementation
Summary: The commit introduces a new ArcanistRectorLinter class which extends the ArcanistBatchExternalLinter class, that aids with PHP upgrades, framework upgrades, and code quality improvements. It handles automated refactoring and returns the results in JSON format. The commit also updates the library map with the reference to the new class.

Test Plan:
[] `arc lint` suggests to fixe files which violate rector rules
[] `arc lint` passes when all violations have been addressed

Reviewers: #php_infrastructure_team, O8 Phabricator!, O43 Developer Experience!

Differential Revision: https://phabricator.tools.flnltd.com/D203519