Add file existence check in getSourceTestFiles function
Summary:
Added a check in the getSourceTestFiles function to verify if the file path exists before it proceeds. This is to prevent potential errors when operating on nonexistent files. The function will now return an empty array if the provided path does not exist.
Ref T298520

Test Plan: [] Remove a src/ class file or a test/unit/src/ file and run `arc unit` it should pass without error (see D205404)

Reviewers: #php_infrastructure_team, O43 Developer Experience!

Subscribers: gmagyar

Maniphest Tasks: T298520

Differential Revision: https://phabricator.tools.flnltd.com/D205451