[flarc] Adjust groovy lint options and flags for new version
Summary:
The new `npm-groovy-lint` version in GAF D209463 requires some changes to inputs/options
https://github.com/nvuillam/npm-groovy-lint/blob/main/CHANGELOG.md

Of note are:
> BREAKING CHANGE:--failon is now ìnfo by default, meaning exit code will be 1 if there is at least an info issue found. To have previous behaviour, use --failon none.
This diff adds a `--failon` option.

> Accept list of directories / files as arguments (--path and --files become deprecated but are still usable)
It says the `--files` option is still usable but when I tested it, it also linted files in `node_modules` which is not what we want. I had to delete the `--files` option to make it work like before which makes it backwards incompatible.

Test Plan:
Tested with D209463, worked for D209456.

I don't know why unit tests are failing but they are evidently unrelated to these changes.

Reviewers: O8 Phabricator, O43 Developer Experience, hzhang

Reviewed By: O8 Phabricator, O43 Developer Experience, hzhang

Maniphest Tasks: T303833

Differential Revision: https://phabricator.tools.flnltd.com/D209459