[flarc] Adjust groovy lint options and flags for new version - take 2
Summary: This is a straight re-diff of D209459, which I reverted because php build is not compatible with this yet (blocked on node 18)

Test Plan:
See D209459 for original tests.

New:
[x] QABox D209463 after node 18 is available in php build.
The overall QABox failed because this diff isn't merged yet, but it got past installing the dependencies.
https://ci.tools.flnltd.com/job/Developers/job/QABox/job/qabox-php-tests/26743/execution/node/210/log/

Reviewers: O8 Phabricator, O43 Developer Experience, fteng, hzhang

Reviewed By: O8 Phabricator, O43 Developer Experience, fteng, hzhang

Maniphest Tasks: T303833

Differential Revision: https://phabricator.tools.flnltd.com/D209481