[flarc] Changed arc lint Stylelint to also read from stderr
Summary:
**Description**
T309695, required for T309374 D212131. Stylelint's v16+ upgrade included a change where errors are properly separated and sent to stderr, while normal stream goes to stdout. An unfortunate byproduct we've encountered is that when formatting output to `json` like we do for `arc lint`, all output goes to stderr, regardless if there's an error or not. {F4362557}

**Solution**
This diff changes the stream the Stylelint linter reads from `$stdout` to `$stderr`. Changing this locally allowed for `arc lint` to pass for v16 upgrade on D212131.

We will probably have to do an announcement of sorts when D212131 is merged in so that people know why their `arc lint` may be suddenly failing, and that they need to update flarc

Test Plan:
I tried running `arc unit` for this repo (on master too) and there's a bunch of unrelated stuff failing
Since this change just affects Stylelint, I'll just verify that the Stylelint unit test passes:
[y] `arc unit .../ArcanistStylelintLinterTestCase.php` passes {F4363751}
[y] On fl-gaf `arc lint` on a Stylelint v16+ branch works {F4363806}
[y] On fl-gaf `arc lint` on a Stylelint v <16 branch works {F4363811} This branch is one I made to try to fix the lint rules before trying again to upgrade v16

Reviewers: O8 Phabricator, O43 Developer Experience, O1 PHP Infra, #developer_experience_team, hzhang

Reviewed By: O8 Phabricator, O43 Developer Experience, O1 PHP Infra, #developer_experience_team, hzhang

Subscribers: bjacobs, hzhang

Maniphest Tasks: T309695

Differential Revision: https://phabricator.tools.flnltd.com/D212171