[rFLARC] Enable specifying interpreter for PHP CodeSniffer
Summary: There's an already existing implementation in `./arcanist/src/lint/linter/ArcanistPhpcsLinter.php` but this implementation doesn't allow us to specify the `interpreter` option in the `.arclint` file. Now, this lightweight wrapper unlocks this possibility to specify the `interpreter` option with a little caveat that we need to change the linter type to `flarc-phpcs`, otherwise an exception will be thrown. **With this change we can run this linter inside docker.**

Test Plan:
[] peer review

[] it successfully runs on rGAF and can find errors

{F4551793 width=100%}

[] it is able to batch files for `phpcs` and runs in one docker container through `bin/gaf-php`

{F4551794 width=100%}

Reviewers: O8 Phabricator, O43 Developer Experience, gread

Reviewed By: O8 Phabricator, O43 Developer Experience, gread

Maniphest Tasks: T318791

Differential Revision: https://phabricator.tools.flnltd.com/D218104